perf(plugin-state): skip redundant namespace quota scans
When the plugin cap is no larger than the namespace cap, its existing
oldest-row eviction also enforces that namespace limit. Avoid the redundant
namespace count and eviction stage for ordinary evict-oldest writes while
preserving the protected key, sibling state, exact live cutoff, and rollback.

Keep reject-new, namespace-only cursor enforcement, Doctor retention, and
expiry cleanup unchanged. Add work-count regression coverage and group the
retention tests with overcapacity, cursor, and rollback checks.
