+++ src/plugin-state/plugin-state-store.test.ts
@@ -3,7 +3,6 @@ import { chmodSync, existsSync, rmSync, statSync } from "node:fs";
 import path from "node:path";
 import { MAX_DATE_TIMESTAMP_MS } from "@openclaw/normalization-core/number-coercion";
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
-import { trackSqliteStatementExecutions } from "../../test/helpers/sqlite-statement-execution-counter.js";
 import { getNodeSqliteKysely } from "../infra/kysely-sync.js";
 import {
   closeOpenClawStateDatabaseByPath,
@@ -23,7 +22,6 @@ import {
   createPluginStateKeyedStore,
   createPluginStateSyncKeyedStore,
   pluginStateEntriesInKeyRange,
-  registerPluginStateSyncSequencedJournalEntry,
   resetPluginStateStoreForTests,
   sweepExpiredPluginStateEntries,
 } from "./plugin-state-store.js";
@@ -202,6 +200,7 @@ describe("plugin state keyed store", () => {
 
   it("updates a key from the current stored value", async () => {
     await withPluginStateTestState(async () => {
+      setMaxPluginStateEntriesPerPluginForTests(10);
       const store = createPluginStateSyncKeyedStore<{ count: number }>("discord", {
         namespace: "sync-update",
         maxEntries: 10,
@@ -341,6 +340,7 @@ describe("plugin state keyed store", () => {
   it("rejects new durable rows at capacity without evicting or blocking updates", async () => {
     await withPluginStateTestState(async () => {
       vi.useFakeTimers();
+      setMaxPluginStateEntriesPerPluginForTests(2);
       const store = createPluginStateKeyedStore<number>("codex", {
         namespace: "durable-bindings",
         maxEntries: 2,
@@ -353,12 +353,16 @@ describe("plugin state keyed store", () => {
 
       await expect(store.register("third", 3)).rejects.toMatchObject({
         code: "PLUGIN_STATE_LIMIT_EXCEEDED",
+        operation: "register",
+        message: "Plugin state namespace durable-bindings for codex reached its 2-row limit.",
       });
       await expect(store.registerIfAbsent("first", 99)).resolves.toBe(false);
       vi.setSystemTime(3000);
       await expect(store.update("first", () => 10)).resolves.toBe(true);
       await expect(store.update("third", () => 3)).rejects.toMatchObject({
         code: "PLUGIN_STATE_LIMIT_EXCEEDED",
+        operation: "register",
+        message: "Plugin state namespace durable-bindings for codex reached its 2-row limit.",
       });
       await expect(store.entries()).resolves.toEqual([
         expect.objectContaining({ key: "second", value: 2 }),
@@ -550,269 +554,6 @@ describe("plugin state keyed store", () => {
       } finally {
         nowSpy.mockRestore();
       }
-    });
-  });
-
-  it("evicts oldest live entries over maxEntries with bounded database calls", async () => {
-    await withPluginStateTestState(async () => {
-      vi.useFakeTimers();
-      vi.setSystemTime(1000);
-      seedPluginStateEntriesForTests(
-        Array.from({ length: 64 }, (_, index) => ({
-          pluginId: "discord",
-          namespace: "evict",
-          key: `key-${String(index).padStart(2, "0")}`,
-          value: index,
-          createdAt: Math.floor(index / 2),
-        })),
-      );
-      const store = createPluginStateKeyedStore("discord", { namespace: "evict", maxEntries: 3 });
-      const statements = trackSqliteStatementExecutions(
-        openOpenClawStateDatabase().db,
-        ["delete"],
-        (sql) => (sql.startsWith('delete from "plugin_state_entries"') ? "delete" : null),
-      );
-      try {
-        await store.register("a-protected", 64);
-      } finally {
-        statements.restore();
-      }
-
-      // One bounded expiry sweep plus eviction must not scale with the victim count.
-      expect(statements.counts.delete).toBeLessThanOrEqual(2);
-      expect(await store.entries()).toEqual([
-        { key: "key-62", value: 62, createdAt: 31 },
-        { key: "key-63", value: 63, createdAt: 31 },
-        { key: "a-protected", value: 64, createdAt: 1000 },
-      ]);
-    });
-  });
-
-  it("keeps the just-registered key when namespace eviction timestamps tie", async () => {
-    await withPluginStateTestState(async () => {
-      vi.useFakeTimers();
-      vi.setSystemTime(1000);
-      const store = createPluginStateKeyedStore<number>("discord", {
-        namespace: "evict-tie-register",
-        maxEntries: 1,
-      });
-
-      await store.register("z", 1);
-      await store.register("a", 2);
-
-      await expect(store.entries()).resolves.toEqual([{ key: "a", value: 2, createdAt: 1000 }]);
-      await expect(store.lookup("z")).resolves.toBeUndefined();
-    });
-  });
-
-  it("keeps a same-millisecond registerIfAbsent claim during namespace eviction", async () => {
-    await withPluginStateTestState(async () => {
-      vi.useFakeTimers();
-      vi.setSystemTime(1000);
-      const store = createPluginStateKeyedStore<number>("discord", {
-        namespace: "evict-tie-claim",
-        maxEntries: 1,
-      });
-
-      await expect(store.registerIfAbsent("z", 1)).resolves.toBe(true);
-      await expect(store.registerIfAbsent("a", 2)).resolves.toBe(true);
-
-      await expect(store.entries()).resolves.toEqual([{ key: "a", value: 2, createdAt: 1000 }]);
-      await expect(store.lookup("z")).resolves.toBeUndefined();
-    });
-  });
-
-  it("evicts current namespace rows when sibling namespaces consume plugin row budget", async () => {
-    await withPluginStateTestState(async () => {
-      const maxPluginEntries = 40;
-      setMaxPluginStateEntriesPerPluginForTests(maxPluginEntries);
-      seedPluginStateEntriesForTests([
-        ...Array.from({ length: maxPluginEntries - 11 }, (_, entryIndex) => ({
-          pluginId: "telegram",
-          namespace: "telegram.message-cache",
-          key: `k-${entryIndex}`,
-          value: { kind: "message", entryIndex },
-        })),
-        ...Array.from({ length: 11 }, (_, entryIndex) => ({
-          pluginId: "telegram",
-          namespace: "telegram.topic-name-cache",
-          key: `topic-${entryIndex}`,
-          value: { kind: "topic", entryIndex },
-        })),
-      ]);
-
-      const messageStore = createPluginStateKeyedStore("telegram", {
-        namespace: "telegram.message-cache",
-        maxEntries: maxPluginEntries,
-      });
-      const topicStore = createPluginStateKeyedStore("telegram", {
-        namespace: "telegram.topic-name-cache",
-        maxEntries: 100,
-      });
-
-      await expect(
-        messageStore.register("new-message", { kind: "message", fresh: true }),
-      ).resolves.toBeUndefined();
-
-      await expect(messageStore.lookup("k-0")).resolves.toBeUndefined();
-      await expect(messageStore.lookup("new-message")).resolves.toEqual({
-        kind: "message",
-        fresh: true,
-      });
-      await expect(topicStore.lookup("topic-0")).resolves.toEqual({
-        kind: "topic",
-        entryIndex: 0,
-      });
-      await expect(messageStore.entries()).resolves.toHaveLength(maxPluginEntries - 11);
-      await expect(topicStore.entries()).resolves.toHaveLength(11);
-    });
-  });
-
-  it("sheds sequenced journal rows without evicting durable sibling state", async () => {
-    await withPluginStateTestState(async () => {
-      const maxPluginEntries = 40;
-      setMaxPluginStateEntriesPerPluginForTests(maxPluginEntries);
-      seedPluginStateEntriesForTests([
-        ...Array.from({ length: maxPluginEntries - 3 }, (_, entryIndex) => ({
-          pluginId: "memory-core",
-          namespace: "durable-state",
-          key: `durable-${entryIndex}`,
-          value: { entryIndex },
-        })),
-        {
-          pluginId: "memory-core",
-          namespace: "memory-host.event-migration-checkpoints",
-          key: "generation",
-          value: { kind: "raw-checkpoint" },
-        },
-        {
-          pluginId: "memory-core",
-          namespace: "memory-host.event-cursors",
-          key: "workspace",
-          value: { kind: "cursor", lastSequence: 1 },
-        },
-        {
-          pluginId: "memory-core",
-          namespace: "memory-host.events",
-          key: "event-1",
-          value: { sequence: 1 },
-        },
-      ]);
-
-      expect(
-        registerPluginStateSyncSequencedJournalEntry({
-          pluginId: "memory-core",
-          cursorOptions: {
-            namespace: "memory-host.event-cursors",
-            maxEntries: 1_000,
-          },
-          cursorKey: "workspace",
-          journalOptions: { namespace: "memory-host.events", maxEntries: 10_000 },
-          initialSequence: 0,
-          journalKey: (sequence) => `event-${sequence}`,
-          journalValue: (sequence) => ({ sequence }),
-        }),
-      ).toBe(2);
-
-      const durable = createPluginStateKeyedStore("memory-core", {
-        namespace: "durable-state",
-        maxEntries: maxPluginEntries,
-      });
-      const checkpoints = createPluginStateKeyedStore("memory-core", {
-        namespace: "memory-host.event-migration-checkpoints",
-        maxEntries: 10_000,
-        overflowPolicy: "reject-new",
-      });
-      const journal = createPluginStateKeyedStore("memory-core", {
-        namespace: "memory-host.events",
-        maxEntries: 10_000,
-      });
-      await expect(durable.lookup("durable-0")).resolves.toEqual({ entryIndex: 0 });
-      await expect(checkpoints.lookup("generation")).resolves.toEqual({
-        kind: "raw-checkpoint",
-      });
-      await expect(journal.lookup("event-1")).resolves.toBeUndefined();
-      await expect(journal.lookup("event-2")).resolves.toEqual({ sequence: 2 });
-    });
-  });
-
-  it("leaves room for Telegram sibling namespaces at their persistent budgets", async () => {
-    await withPluginStateTestState(async () => {
-      seedPluginStateEntriesForTests([
-        ...Array.from({ length: 3_000 }, (_, entryIndex) => ({
-          pluginId: "telegram",
-          namespace: "telegram.message-cache",
-          key: `message-${entryIndex}`,
-          value: { kind: "message", entryIndex },
-        })),
-        ...Array.from({ length: 2_047 }, (_, entryIndex) => ({
-          pluginId: "telegram",
-          namespace: "telegram.topic-name-cache",
-          key: `topic-${entryIndex}`,
-          value: { kind: "topic", updatedAt: entryIndex },
-        })),
-        ...Array.from({ length: 127 }, (_, entryIndex) => ({
-          pluginId: "telegram",
-          namespace: "telegram.bot-info-cache",
-          key: `bot-${entryIndex}`,
-          value: { kind: "bot-info", fetchedAt: String(entryIndex) },
-        })),
-      ]);
-
-      const topicStore = createPluginStateKeyedStore("telegram", {
-        namespace: "telegram.topic-name-cache",
-        maxEntries: 2_048,
-      });
-      const botInfoStore = createPluginStateKeyedStore("telegram", {
-        namespace: "telegram.bot-info-cache",
-        maxEntries: 128,
-      });
-
-      await expect(
-        topicStore.register("topic-final", { kind: "topic", updatedAt: 2_048 }),
-      ).resolves.toBeUndefined();
-      await expect(
-        botInfoStore.register("default", { kind: "bot-info", fetchedAt: "now" }),
-      ).resolves.toBeUndefined();
-
-      await expect(topicStore.lookup("topic-final")).resolves.toEqual({
-        kind: "topic",
-        updatedAt: 2_048,
-      });
-      await expect(botInfoStore.lookup("default")).resolves.toEqual({
-        kind: "bot-info",
-        fetchedAt: "now",
-      });
-    });
-  });
-
-  it("rejects plugin overflow when the current namespace cannot shed old rows", async () => {
-    await withPluginStateTestState(async () => {
-      const maxPluginEntries = 40;
-      setMaxPluginStateEntriesPerPluginForTests(maxPluginEntries);
-      seedPluginStateEntriesForTests(
-        Array.from({ length: maxPluginEntries }, (_, entryIndex) => ({
-          pluginId: "telegram",
-          namespace: "telegram.topic-name-cache",
-          key: `topic-${entryIndex}`,
-          value: { entryIndex },
-        })),
-      );
-
-      const messageStore = createPluginStateKeyedStore("telegram", {
-        namespace: "telegram.message-cache",
-        maxEntries: maxPluginEntries,
-      });
-      const topicStore = createPluginStateKeyedStore("telegram", {
-        namespace: "telegram.topic-name-cache",
-        maxEntries: maxPluginEntries,
-      });
-
-      await expectPluginStateStoreError(messageStore.register("new-message", { fresh: true }), {
-        code: "PLUGIN_STATE_LIMIT_EXCEEDED",
-      });
-      await expect(messageStore.lookup("new-message")).resolves.toBeUndefined();
-      await expect(topicStore.lookup("topic-0")).resolves.toEqual({ entryIndex: 0 });
     });
   });
